requirements.txt
Flask==2.0.3
flask-cors
gevent==21.12.0
keras==2.9.0
numpy==1.21.4
opencv_python_headless==4.6.0.66
Pillow==9.4.0
Werkzeug==2.0.2
tensorflow ==2.9.1